Refactor: extracted traits impls.
